+++ CHANGELOG.md
@@ -38,6 +38,8 @@ According to [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) , the
   All Rate Policy Entries for an Installation.
 - [#294](https://github.com/ARGOeu/argo-accounting/pull/294) - ACC-1875 PATCH
   Rate for a Single Metric Definition on an Installation.
+- [#295](https://github.com/ARGOeu/argo-accounting/pull/295) - ACC-1869 Credit
+  Usage Report Across Installations for a Time Range.
 
 ## 1.8.1 - 2025-12-09